+++ backend/lost/api/data/endpoint.py
@@ -12,6 +12,7 @@ from lost.logic import anno_task as annotask_service
 from lost.logic.file_man import FileMan
 from lost.logic import dask_session
 from lost.pyapi.utils import anno_helper
+from lost.logic.file_access import UserFileAccess
 import json
 import os
 from lost.pyapi import pe_base
@@ -141,12 +142,16 @@ class Logs(Resource):
             # return send_from_directory(os.path.join(LOST_CONFIG.project_path, 'logs'), path)
 
 
-def load_img(db_img, fm, user):
+def load_img(db_img, ufa, user):
     if LOST_CONFIG.worker_management != 'dynamic':
         # need to execute ls for s3fs (don't know why)
-        fm.fs.ls(db_img.img_path)
-        img = fm.load_img(db_img.img_path)
+        try:
+            ufa.fs.ls(db_img.img_path)
+        except:
+            pass
+        img = ufa.load_anno_img(db_img)
     else:
+        # TODO: Use UserFileAccess to load image
         img = dask_session.ds_man.read_fs_img(user, db_img.fs, db_img.img_path)
     return img
 
@@ -167,15 +172,16 @@ class GetImage(Resource):
             data = json.loads(request.data)
             #flask.current_app.logger.info('mia -> getimage. Received data: {}'.format(data))
             if data['type'] == 'imageBased':
+                # TODO: get image via db access
                 db_img = dbm.get_image_anno(data['id'])
-                fm = FileMan(fs_db=db_img.fs) 
-                img = load_img(db_img, fm, user)
+                ufa = UserFileAccess(dbm, user.idx, db_img.fs)
+                img = load_img(db_img, ufa, user)
             elif data['type'] == 'annoBased':
                 db_anno = dbm.get_two_d_anno(two_d_anno_id=data['id'])
                 db_img = dbm.get_image_anno(db_anno.img_anno_id)
-                fm = FileMan(fs_db=db_img.fs)
+                ufa = UserFileAccess(dbm, user.idx, db_img.fs)
                 # image = fm.load_img(db_img.img_path)
-                image = load_img(db_img, fm, user)
+                image = load_img(db_img, ufa, user)
                 
                 # get annotation_task config
                 draw_anno = False